@@ -38,9 +31,10 @@
     "@material-ui/icons": "4.11.2",
     "@material-ui/styles": "4.11.4",
     "cheerio": "1.0.0-rc.10",
+    "next": "12.1.5",
     "promise-polyfill": "^8.1.0",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "18.0.0",
+    "react-dom": "18.0.0",
     "react-ga": "3.3.0",
     "react-plx": "^1.3.11",
     "react-scrollspy": "^3.3.5",
