@@ -9,7 +9,7 @@ import IconButton from "@material-ui/core/IconButton";
 import Brightness3Icon from "@material-ui/icons/Brightness3";
 import BrightnessHighIcon from "@material-ui/icons/BrightnessHigh";
 import MenuIcon from "@material-ui/icons/Menu";
-import { useDarkLightModeToggler } from "services/MuiThemeFactory";
+import { useDarkLightModeToggler } from "@/src/services/MuiThemeFactory";
 
 import OldVersionLinkButton from "./OldVersionLinkButton";
 
@@ -55,7 +55,7 @@ export default function PageAppBar({ onIconButtonClick }: PageAppBarProps) {
                     <MenuIcon />
                 </IconButton>
                 <Typography variant="h6" color="inherit" noWrap className={classNames(classes.title)}>
-                    Hey's Personal Portfolio
+                    Hey&apos;s Personal Portfolio
                 </Typography>
                 <>
                     <IconButton aria-label="dark-mode" onClick={toggleDarkLightMode}>
